refactor: Page 컴포넌트도 components 아래로 이동
