+++ src/pages/MagicBook/components/AnswerPage.tsx
@@ -1,5 +1,5 @@
 import styles from '@/pages/MagicBook/MagicBook.module.css';
-import Page from '@/pages/MagicBook/Page.tsx';
+import Page from '@/pages/MagicBook/components/Page.tsx';
 import { PageIndex } from '@/pages/MagicBook/types.ts';
 
 function AnswerPage({ index }: Readonly<PageIndex>) {
